schema 모델들 분류해서 정리
