+++ app/schemas/__init__.py
@@ -1,0 +1,4 @@
+from .quiz import *
+from .user import *
+from .scoreboard import *
+from .community import *
