+++ app/schemas/quiz.py
@@ -1,0 +1,11 @@
+from typing import List
+
+from pydantic import BaseModel
+
+
+class RandomQuizResponse(BaseModel):
+    path: str
+
+
+class TestQuizResponse(BaseModel):
+    order: List[str]
