+++ app/schemas/user.py
@@ -1,0 +1,19 @@
+from datetime import datetime
+
+from pydantic import BaseModel
+
+
+class UserBase(BaseModel):
+    email: str
+
+
+class UserCreate(UserBase):
+    pass
+
+
+class User(UserBase):
+    id: int
+    created_at: datetime
+
+    class Config:
+        orm_mode = True
